Bug 2011353 - Change CurrentTimeFromTimelineTime and StartTimeFromTimelineTime's type of playback rate. r=boris
These functions were introduced in this commit [1], the reason why the
type was float is probably this StartTimeFromTimelineTime call site [2]
where the playback rate comes from this playbackRate variable [3] for
compositor animations, which is defined as float. It doesn't make sense
to reduce accuracy for the single call site.

[1] https://hg-edge.mozilla.org/mozilla-central/rev/e11bc080108c4362a2f2afc828f4a1cfc7cd33c2
[2] https://searchfox.org/firefox-main/rev/08a5a0de94770126c13dceb661fee2edbdff0329/gfx/layers/AnimationInfo.cpp#105-106
[3] https://searchfox.org/firefox-main/rev/08a5a0de94770126c13dceb661fee2edbdff0329/gfx/layers/ipc/LayersMessages.ipdlh#214

Differential Revision: https://phabricator.services.mozilla.com/D282972

+++ dom/animation/Animation.h
@@ -221,7 +221,7 @@ class Animation : public DOMEventTargetHelper,
    */
   static TimeDuration CurrentTimeFromTimelineTime(
       const TimeDuration& aTimelineTime, const TimeDuration& aStartTime,
-      float aPlaybackRate) {
+      double aPlaybackRate) {
     return (aTimelineTime - aStartTime).MultDouble(aPlaybackRate);
   }
 
@@ -236,7 +236,7 @@ class Animation : public DOMEventTargetHelper,
    */
   static TimeDuration StartTimeFromTimelineTime(
       const TimeDuration& aTimelineTime, const TimeDuration& aCurrentTime,
-      float aPlaybackRate) {
+      double aPlaybackRate) {
     TimeDuration result = aTimelineTime;
     if (aPlaybackRate == 0) {
       return result;
